chore(masonry): move new brick utils to src/utils
