+++ modules/masonry/src/brick/view/components/Path.tsx
@@ -1,7 +1,7 @@
 import type { BrickOutlineInput } from '@/@types/brick';
 
-import { SCALE_LEVEL_CONFIG } from '../../utils/constants';
-import { BrickOutlineGenerator } from '../../utils/path2';
+import { SCALE_LEVEL_CONFIG } from '@/utils/constants';
+import { BrickOutlineGenerator } from '@/utils/path';
 
 const SCALE_LEVEL: keyof typeof SCALE_LEVEL_CONFIG = 2;
 const SCALE = SCALE_LEVEL_CONFIG[SCALE_LEVEL].brickScale;
